Added luminance normalization option

+++ scripts/bbbc021-dcgan-profile.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-rm dcgan.prof
+if [[ -f dcgan.prof ]]; then
+  rm dcgan.prof
+fi
 
 python3 -m cProfile -o dcgan.prof cytogan/train/bbbc021.py \
   --epochs 30                                              \
@@ -14,7 +16,7 @@ python3 -m cProfile -o dcgan.prof cytogan/train/bbbc021.py \
   --summary-freq '1min'                                    \
   --latent-samples 256                                     \
   --generative-samples 5                                   \
-  --gpus 1 2                                               \
+  --gpus 2 3                                               \
   --metadata /data1/peter/metadata/BBBC021_v1_image.csv    \
   --labels /data1/peter/metadata/BBBC021_v1_moa.csv        \
   --images /data1/peter/segmented                          \
